+++ Bet apps/src/chorke/proprietary/bet/apps/core/bets/BetDrawNoBet.java
@@ -4,15 +4,17 @@
  */
 package chorke.proprietary.bet.apps.core.bets;
 
+import java.math.BigDecimal;
+
 /**
  *
  * @author Chorke
  */
 public class BetDrawNoBet extends Bet{
-    public final double bet1;
-    public final double bet2;
+    public final BigDecimal bet1;
+    public final BigDecimal bet2;
 
-    public BetDrawNoBet(String betCompany, double bet1, double bet2) {
+    public BetDrawNoBet(String betCompany, BigDecimal bet1, BigDecimal bet2) {
         super(betCompany);
         this.bet1 = bet1;
         this.bet2 = bet2;
